% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{clean_input}
\alias{clean_input}
\title{clean_input}
\usage{
clean_input(x)
}
\arguments{
\item{x}{an input value}
}
\value{
NA if input is "" or NULL or "missing", otherwise returns x
}
\description{
makes blank input values ("") and NULL and "missing" into NAs
}
