% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{dummy_model}
\alias{dummy_model}
\title{Dummy model fitted based on a standard GAM}
\usage{
dummy_model(df_loaded, predgrid, n_post_samples, n_chains)
}
\arguments{
\item{df_loaded}{a `data.frame` of data}

\item{predgrid}{prediction grid to use}

\item{n_post_samples}{how many posterior samples to use}

\item{n_chains}{number of chains (just a multiplier on the number of samples
here but when we do Real MCMC we will need use this)}
}
\value{
df_results
}
\description{
Nothing fancy here! Just a simple GAM, followed by some uncertainty
estimation using posterior sampling. More information at
https://github.com/NERC-CEH/soccatoa_app/issues/4
}
