% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{format_model_data}
\alias{format_model_data}
\title{Process data for modelling}
\usage{
format_model_data(df_loaded)
}
\arguments{
\item{df_loaded}{the data as loaded}
}
\value{
a data.frame that is ready for modelling
}
\description{
Maybe want to put this elsewhere, but a little bit of data fiddling.
}
\details{
What this function does:
 - assumes lat/lon columns, EPSG4326
 - project to UK grid EPSG27700
 - creates new columns called easting and northing
 - creates rho_c = f_c*rho_fe
 - calculates log rho_c
 - makes a factor year variable
}
