% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{get_co2climate_effect}
\alias{get_co2climate_effect}
\title{Calculate the change in soil carbon and its uncertainty arising from change
in CO2 and temperature.}
\usage{
get_co2climate_effect(
  n_sim = 10,
  S_c_start = 10,
  start_year = 2025,
  end_year = 2050,
  this_scenario = c("RCP2.6", "RCP4.5", "RCP6.0", "RCP8.5"),
  shape = 3.920758,
  rate = 1371.239,
  intercept = 0.004518087,
  slope = -3.611591834,
  sigma = 0.007210343,
  maxbeta
)
}
\arguments{
\item{n_sim}{The number of simulations to run (default: 10)}

\item{S_c_start}{The initial soil carbon stock (default: 10 kg C m^-2)}

\item{start_year}{The starting year for the calculation (default: 2025)}

\item{end_year}{The ending year for the calculation (default: 2050)}

\item{this_scenario}{The climate scenario to use (default: "RCP2.6")}

\item{shape}{The shape parameter for the gamma distribution (default: 3.920758)}

\item{rate}{The rate parameter for the gamma distribution (default: 1371.239)}

\item{intercept}{The intercept for the linear relationship between beta and gamma (default: 0.004518087)}

\item{slope}{The slope for the linear relationship between beta and gamma (default: -3.611591834)}

\item{sigma}{The standard deviation for the normal distribution of residuals (default: 0.007210343)}
}
\value{
A list with two elements:
  \item{sigma}{The standard deviation of the estimated soil carbon stock in the end year}
  \item{p}{A ggplot object showing the distribution of predicted soil carbon stocks}
}
\description{
This function calculates the change in soil carbon stock from the start year
to the end year, and its standard deviation, given the parameters
describing the sensitivity of SOC to change in CO2 and temperature.
}
\examples{
\dontrun{
x <- get_co2climate_effect(
  n_sim = 10000,
  start_year = 2025,
  end_year = 2050,
  this_scenario = "RCP8.5"
)
}
}
