% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{get_right_date_format}
\alias{get_right_date_format}
\title{transform date}
\usage{
get_right_date_format(x, type)
}
\arguments{
\item{x}{column values}

\item{type}{starting date format}
}
\value{
the column values in date format d/%m/%Y
}
\description{
function to transfrom column into right date format
}
