% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{get_standard_unit}
\alias{get_standard_unit}
\title{get_standard_unit}
\usage{
get_standard_unit(my_columns, my_columns_unit, loaded_data)
}
\arguments{
\item{my_columns}{column names in uploaded data}

\item{my_columns_unit}{columns unit of data uplooaded}

\item{loaded_data}{data loaded by user}
}
\value{
a df with the variables in the right unit
}
\description{
function to transfrom the provided dat into t he DB unit
}
