% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{get_unit_choices}
\alias{get_unit_choices}
\title{get_unit_choices for each variable}
\usage{
get_unit_choices(var_name)
}
\arguments{
\item{var_name}{id var name to which get units options for}
}
\value{
the choices to display as vector
}
\description{
function to get which units to show for which variables
}
