% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{latlon_to_ne}
\alias{latlon_to_ne}
\title{Lat/lon to northing/easting}
\usage{
latlon_to_ne(df, crs = 27700)
}
\arguments{
\item{df}{the data as initially uploaded}

\item{crs}{the coordinate reference system to project into (can be any
format accepted by [sf::st_transform]}
}
\value{
a data.frame that is ready for modelling
}
\description{
A little bit of data fiddling to project from lat/lon columns, EPSG4326 to
(by default) UK grid EPSG27700. This is for the model to use.
}
