% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{make_prediction_grid}
\alias{make_prediction_grid}
\title{Make a prediction grid}
\usage{
make_prediction_grid(df_loaded)
}
\arguments{
\item{...}{nothing for now but we might want options in the future}
}
\value{
a `data.frame` with locations in covariate space to predict at
}
\description{
Generate a grid to make predictions over
}
