% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{ne_to_latlon}
\alias{ne_to_latlon}
\title{Northing/easting to lat/lon}
\usage{
ne_to_latlon(df)
}
\arguments{
\item{df}{the data as initially uploaded}
}
\value{
a data.frame that is ready for modelling
}
\description{
A little bit of data fiddling to project from UK grid EPSG27700 to lat/lon
columns, EPSG4326. This is for storage in the DB.
}
