% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-others.R
\name{render_select}
\alias{render_select}
\title{Generate HTML select tags for table}
\usage{
render_select(id, choices)
}
\arguments{
\item{id}{id to use in the select input a character}

\item{choices}{choices to add as vector}
}
\value{
string that creates select input of id id and choices choices in HTML
}
\description{
function ...
}
