% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{rgammat}
\alias{rgammat}
\title{Random Gamma Distribution within a specified range}
\usage{
rgammat(n, range, shape, rate = 1)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{range}{A numeric vector of length 2 specifying the minimum and maximum values of the range.}

\item{shape}{Shape parameter of the Gamma distribution.}

\item{rate}{Rate parameter of the Gamma distribution. Defaults to 1.}
}
\value{
A numeric vector of length `n` containing random numbers from a Gamma distribution within the specified range.
}
\description{
Generates random numbers from a Gamma distribution within a specified range.
}
\examples{
set.seed(123)
rgammat(10, range = c(0, 10), shape = 2, rate = 1)
}
