% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{run_model_A}
\alias{run_model_A}
\title{Spatio-temporal model of soil carbon}
\usage{
run_model_A(df, start_year, end_year)
}
\arguments{
\item{df}{a `data.frame` of data loaded}
}
\value{
df_results
}
\description{
This fits our model of soil carbon. At the moment this is just a dummy
function and will eventually call-out to other stuff.
}
