% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{summarize_results_dist}
\alias{summarize_results_dist}
\title{Summarize results for uncertainty plot}
\usage{
summarize_results_dist(results)
}
\arguments{
\item{results}{a result object from [run_model_A]}

\item{a}{`data.frame` with columns:}
}
\description{
Calculates the posterior of the change in carbon
}
