% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run_model.R
\name{run_model_B}
\alias{run_model_B}
\title{run_model_B}
\usage{
run_model_B(yrstart, yrend)
}
\arguments{
\item{yrstart}{start year}

\item{yrend}{end year}
}
\value{
df_results
}
\description{
function ...
}
